LuxuryMarket/core/serializers.py
# core/serializers.py
from rest_framework import serializers
from .models import Listing, UserProfile

class ListingSerializer(serializers.ModelSerializer):
    class Meta:
        model = Listing
        fields = '__all__'
        
class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = '__all__'